refactor: rename utils dir to validators dir
