+++ backend/backend/validators/drf_validators.py
@@ -4,7 +4,7 @@ import phonenumbers
 from django.utils.translation import gettext_lazy as _
 from rest_framework import serializers
 
-from backend.utils.validation import is_valid_phone_number
+from backend.validators.validation import is_valid_phone_number
 
 
 def validate_phone_number(value: str) -> str:
